Move ByteBufVisitor to org.apache.bookkeeper.util package
